+++ lib/commands/runner.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 import 'package:args/command_runner.dart';
 import 'package:gd/commands/create_command.dart';
 import 'package:gd/commands/install_command.dart';
+import 'package:gd/commands/linux/linux_runner.dart';
 import 'package:gd/commands/windows/windows_runner.dart';
 import 'package:gd/ui/core_ui.dart';
 
@@ -24,6 +25,8 @@ Future<CommandRunner> createRunner(final CoreUi ui) {
     ..addCommand(CreateCommand(ui));
   if (Platform.isWindows) {
     return createWindowsRunner(ui, runner);
+  } else if (Platform.isLinux) {
+    return createLinuxRunner(ui, runner);
   }
   throw UnimplementedError();
 }